// Quellcode/src/app/Model/arbeitslose.ts
export class Arbeitslose{
  id: number | null = null
  datum: string= "";
  col1: number | null = null;
  col2: number | null = null;
  col3: number | null = null;
  col4: number | null = null;
  col5: number | null = null;
  col6: number | null = null;
  col7: number | null = null;
  col8: number | null = null;
  col9: number | null = null;
  col10: number | null = null;
  col11: number | null = null;
  col12: number | null = null;
}